Update readme, clean up code
