+++ Example/SwiftUIKitExample/SwiftUIwithUIKitView.swift
@@ -14,20 +14,13 @@ struct SwiftUIwithUIKitView: View {
     var body: some View {
         NavigationView {
             VStack {
-//                UIViewMaker<UILabel> {
-//                    $0.text = "Hello no \(self.integer) from UIKit"
-//                }
-//                .fixedSize()
-//                UIViewContainer(UILabel(), layout: .intrinsic)
-//                    .set(\.text, to: "Hello, UIKit \(integer)!")
-//                    .set(\.backgroundColor, to: UIColor(named: "swiftlee_orange"))
-//                    .fixedSize()
+                // Use UIKit inside SwiftUI like this:
                 UIViewContainer(UIKitView(), layout: .intrinsic)
                     .set(\.title, to: "Hello, UIKit \(integer)!")
                     .set(\.backgroundColor, to: UIColor(named: "swiftlee_orange"))
                     .fixedSize()
-//                    .fixedSize()
                     .navigationTitle("Use UIKit in SwiftUI")
+
                 Button("RANDOMIZED: \(integer)") {
                     integer = Int.random(in: 0..<300)
                 }
@@ -53,52 +46,3 @@ struct UILabelExample_Preview: PreviewProvider {
             .previewDisplayName("UILabel Preview Example")
     }
 }
-
-struct UIViewMaker<ViewType: UIView>: UIViewRepresentable {
-
-    typealias UIViewType = ViewType
-
-    var make: () -> ViewType = ViewType.init
-    var update: (ViewType) -> ()
-
-    func makeUIView(context: Context) -> ViewType {
-        return make()
-    }
-
-    func updateUIView(_ uiView: ViewType, context: Context) {
-        update(uiView)
-    }
-}
-
-struct ContentView: View {
-    @State var counter = 0
-
-    var body: some View {
-        VStack {
-
-            Text("Hello no \(counter) from SwiftUI")
-                .padding()
-
-
-            UIViewMaker<UILabel> {
-                $0.text = "Hello no \(self.counter) from UIKit"
-            }
-            .fixedSize()
-
-        }
-        .onAppear {
-            if counter == 0 {
-                schedule()
-            }
-        }
-    }
-
-    func schedule() {
-        counter += 1
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-            self.schedule()
-        }
-    }
-
-}
-
